Update underlying records drill - bring to top, label changes (#29984)
