+++ docs/exploration-and-organization/exploration.md
@@ -43,7 +43,7 @@ If you click on a part of a chart, such as a bar in a bar chart, or a dot on a l
 In this example of orders by product category over time, clicking on a dot on this line chart gives us the ability to:
 
 - **Zoom in**: See orders for a particular category over a shorter time range.
-- **View these Orders**: See a list of the orders for a particular month
+- **See these Orders**: See a list of the orders for a particular month
 - **Break out by a category**: See things like the Gizmo orders in June 2017 broken out by the status of the customer (e.g., `new` or `VIP`). Different charts will have different breakout options, such as **Location** and **Time**.
 
 > Note that while charts created with SQL don't currently have the drill-through menu, you can add SQL questions to a dashboard and customize their click behavior. You can send people to a [custom destination](https://www.metabase.com/learn/building-analytics/dashboards/custom-destinations.html) (like another dashboard or an external URL), or have the clicked value [update a dashboard filter](https://www.metabase.com/learn/building-analytics/dashboards/cross-filtering.html).
