+++ e2e/test/scenarios/filters/reproductions/18770-post-aggregation-filter-disrupts-drillthrough.cy.spec.js
@@ -40,7 +40,7 @@ describe.skip("issue 18770", () => {
 
     cy.findByText("4,784").click();
     popover()
-      .should("contain", "View these Orders")
+      .should("contain", "See these Orders")
       .and("contain", "Break out by a…")
       .and("contain", "Filter by this value")
       .and("contain", "Automatic explorations");
