+++ e2e/test/scenarios/models/reproductions/29517-native-remapped-model-drill-through-click-behavior.cy.spec.js
@@ -82,7 +82,7 @@ describe("issue 29517 - nested question based on native model with remapped valu
     // We can click on any circle; this index was chosen randomly
     cy.get("circle").eq(25).click({ force: true });
     popover()
-      .findByText(/^View these/)
+      .findByText(/^See these/)
       .click();
     cy.wait("@dataset");
