+++ e2e/test/scenarios/permissions/sandboxes.cy.spec.js
@@ -335,7 +335,7 @@ describeEE("formatting > sandboxes", () => {
           // Click on the first bar in a graph (Category: "Doohickey")
           cy.get(".bar").eq(0).click({ force: true });
         });
-        cy.findByText("View these Orders").click();
+        cy.findByText("See these Orders").click();
 
         cy.log("Reported failing on v1.37.0.2");
         cy.wait("@dataset").then(xhr => {
@@ -408,7 +408,7 @@ describeEE("formatting > sandboxes", () => {
         // Click on the first bar in a graph (Category: "Doohickey")
         cy.get(".bar").eq(0).click({ force: true });
       });
-      cy.findByText("View these Orders").click();
+      cy.findByText("See these Orders").click();
 
       cy.wait("@dataset");
       cy.log("Reported failing on v1.36.4");
@@ -694,7 +694,7 @@ describeEE("formatting > sandboxes", () => {
           // Click on the second bar in a graph (Category: "Widget")
           cy.get(".bar").eq(1).click({ force: true });
         });
-        cy.findByText("View these Orders").click();
+        cy.findByText("See these Orders").click();
 
         cy.wait("@dataset").then(xhr => {
           expect(xhr.response.body.error).not.to.exist;
