+++ frontend/src/metabase/modes/components/drill/UnderlyingRecordsDrill.ts
@@ -6,6 +6,8 @@ import {
 } from "metabase-lib/queries/drills/underlying-records-drill";
 import type { Drill } from "../../types";
 
+const isShortTableName = (tableName: string) => tableName.length <= 20;
+
 export const UnderlyingRecordsDrill: Drill = ({ question, clicked }) => {
   const drill = underlyingRecordsDrill({ question, clicked });
   if (!drill) {
@@ -14,13 +16,14 @@ export const UnderlyingRecordsDrill: Drill = ({ question, clicked }) => {
 
   const { tableName, rowCount } = drill;
 
-  const tableTitle = tableName
-    ? inflect(tableName, rowCount)
-    : ngettext(msgid`record`, `records`, rowCount);
+  const tableTitle =
+    tableName && isShortTableName(tableName)
+      ? inflect(tableName, rowCount)
+      : ngettext(msgid`record`, `records`, rowCount);
 
   const actionTitle = ngettext(
-    msgid`View this ${tableTitle}`,
-    `View these ${tableTitle}`,
+    msgid`See this ${tableTitle}`,
+    `See these ${tableTitle}`,
     rowCount,
   );
