+++ frontend/src/metabase/modes/types.ts
@@ -18,15 +18,15 @@ export type ClickObject = {
   value?: RowValue;
   column?: DatasetColumn;
   dimensions?: DimensionValue[];
-  settings?: Record<string, unknown>;
-  extraData?: Record<string, unknown>;
+  event?: MouseEvent;
+  element?: HTMLElement;
   seriesIndex?: number;
+  settings?: Record<string, unknown>;
   origin?: {
     row: RowValue;
     cols: DatasetColumn[];
   };
-  event?: MouseEvent;
-  element?: HTMLElement;
+  extraData?: Record<string, unknown>;
 };
 
 type Dispatcher = (dispatch: Dispatch) => void;
@@ -89,6 +89,7 @@ type AlwaysDefaultClickAction = Omit<
   defaultAlways: true;
 };
 
+// TODO [#26836]: unify this and frontend/src/metabase-types/types/Visualization.ts
 export type ClickAction = RegularClickAction | AlwaysDefaultClickAction;
 
 export type DrillOptions = {
