+++ frontend/src/metabase/visualizations/components/ChartClickActions/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ChartClickActions";
